Crate rusoto_kinesisvideo

If you're using the service, you're probably looking for KinesisVideoClient and KinesisVideo.

Structs

CreateStreamInput

CreateStreamOutput

DeleteStreamInput

DeleteStreamOutput

DescribeStreamInput

DescribeStreamOutput

GetDataEndpointInput

GetDataEndpointOutput

KinesisVideoClient
A client for the Kinesis Video API.

ListStreamsInput

ListStreamsOutput

ListTagsForStreamInput

ListTagsForStreamOutput

StreamInfo
An object describing a Kinesis video stream.

StreamNameCondition
Specifies the condition that streams must satisfy to be returned when you list streams (see the ListStreams API). A condition has a comparison operation and a value. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

TagStreamInput

TagStreamOutput

UntagStreamInput

UntagStreamOutput

UpdateDataRetentionInput

UpdateDataRetentionOutput

UpdateStreamInput

UpdateStreamOutput

Enums

CreateStreamError
Errors returned by CreateStream

DeleteStreamError
Errors returned by DeleteStream

DescribeStreamError
Errors returned by DescribeStream

GetDataEndpointError
Errors returned by GetDataEndpoint

ListStreamsError
Errors returned by ListStreams

ListTagsForStreamError
Errors returned by ListTagsForStream

TagStreamError
Errors returned by TagStream

UntagStreamError
Errors returned by UntagStream

UpdateDataRetentionError
Errors returned by UpdateDataRetention

UpdateStreamError
Errors returned by UpdateStream

Traits

KinesisVideo
Trait representing the capabilities of the Kinesis Video API. Kinesis Video clients implement this trait.