Struct rusoto_kinesisvideo::KinesisVideoClient

pub struct KinesisVideoClient { /* fields omitted */ }

A client for the Kinesis Video API.

Methods

impl KinesisVideoClient

pub fn new(region: Region) -> KinesisVideoClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> KinesisVideoClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,